todoRedy

+++ routes/web.php
@@ -2,6 +2,7 @@
 
 use App\Http\Controllers\ArchivosController;
 use App\Http\Controllers\CalendarioController;
+use App\Http\Controllers\CartController;
 use App\Http\Controllers\ColposcopiasController;
 use Illuminate\Support\Facades\Route;
 use Laravel\Fortify\Features;
@@ -13,6 +14,8 @@ use App\Http\Controllers\UserController;
 use App\Http\Controllers\PacientesController;
 use App\Http\Controllers\ConsultasController;
 use App\Http\Controllers\DiagnosticosController;
+use App\Http\Controllers\HojaHospitalController;
+use App\Http\Controllers\HospitalizacionController;
 use App\Http\Controllers\NotasController;
 use App\Http\Controllers\TriajesController;
 
@@ -50,7 +53,8 @@ Route::middleware([
 
     Route::group(['prefix' => 'calendario'], function () {
         Route::get('/', [CalendarioController::class, 'index'])->name('calendario');
-        Route::post('{id}', [CalendarioController::class, 'store'])->name('calendario.store');
+        Route::post('{id}', [CalendarioController::class, 'store'])->middleware(['role:Administrador,Enfermeria consultorios,Medico especialista'])
+        ->name('calendario.store');
     });
 
     Route::group(['prefix' => 'gestion-usuarios', 'middleware' => ['role:Administrador']], function () {
@@ -59,12 +63,12 @@ Route::middleware([
         Route::delete('{id}', [UserController::class, 'destroy'])->name('usuarios.eliminar');
     });
 
-    Route::group(['prefix' => 'pacientes', 'middleware' => ['role:Administrador,Enfermeria consultorios,Medico especialista,Medico general,Enfermeria hospitalizacion']], function () {
-        Route::get('/', [PacientesController::class, 'index'])->name('pacientes');
-        Route::get('{id}', [PacientesController::class, 'show'])->name('pacientes.show');
-        Route::post('{id}', [PacientesController::class, 'store'])->name('pacientes.store');
-        Route::put('{id}', [PacientesController::class, 'update'])->name('pacientes.update');
-        Route::delete('{id}', [PacientesController::class, 'destroy'])->name('pacientes.eliminar');
+    Route::group(['prefix' => 'pacientes'], function () {
+        Route::get('/', [PacientesController::class, 'index'])->middleware(['role:Administrador,Enfermeria consultorios,Medico especialista,Medico general,Enfermeria hospitalizacion'])->name('pacientes');
+        Route::get('{id}', [PacientesController::class, 'show'])->middleware(['role:Administrador,Enfermeria consultorios,Medico especialista,Medico general,Enfermeria hospitalizacion'])->name('pacientes.show');
+        Route::post('{id}', [PacientesController::class, 'store'])->middleware(['role:Administrador,Enfermeria consultorios,Medico especialista'])->name('pacientes.store');
+        Route::put('{id}', [PacientesController::class, 'update'])->middleware(['role:Administrador,Enfermeria consultorios,Medico especialista'])->name('pacientes.update');
+        Route::delete('{id}', [PacientesController::class, 'destroy'])->middleware(['role:Administrador,Enfermeria consultorios,Medico especialista'])->name('pacientes.eliminar');
     });
 
     Route::group(['prefix' => 'consultas'], function () {
@@ -84,10 +88,10 @@ Route::middleware([
             ->middleware(['role:Administrador,Enfermeria consultorios,Medico especialista'])->name('consultas.eliminar');
 
         Route::get('consultas_dia', [ConsultasController::class, 'consultas_dia'])
-            ->middleware(['role:Administrador,Enfermeria consultorios,Medico especialista'])->name('consultas_dia');
+            ->middleware(['role:Administrador,Enfermeria consultorios,Medico especialista,Medico general'])->name('consultas_dia');
 
         Route::get('consultas_espera', [ConsultasController::class, 'consultas_espera'])
-            ->middleware(['role:Administrador,Enfermeria consultorios,Medico especialista'])->name('consultas_espera');
+            ->middleware(['role:Administrador,Enfermeria consultorios,Medico especialista,Medico general'])->name('consultas_espera');
     });
 
     Route::group(['prefix' => '/', 'middleware' => ['role:Administrador,Enfermeria consultorios,Medico especialista,Medico general,Enfermeria hospitalizacion']], function () {
@@ -125,6 +129,38 @@ Route::middleware([
         Route::post('{id}', [NotasController::class, 'store'])->name('nota.store');
         Route::put('{id}', [NotasController::class, 'update'])->name('nota.update');
     });
+
+    Route::group(['prefix' => 'hospitalizacion'], function () {
+        Route::get('/', [HospitalizacionController::class, 'index'])
+            ->middleware(['role:Administrador,Medico general,Medico especialista,Enfermeria hospitalizacion,Contador'])->name('hospitalizacion');    
+        Route::get('show/{id}', [HospitalizacionController::class, 'show'])
+            ->middleware(['role:Administrador,Medico general,Medico especialista,Enfermeria hospitalizacion,Contador'])->name('hospitalizacion.show');
+        Route::get('{id}', [HospitalizacionController::class, 'create'])
+            ->middleware(['role:Administrador,Medico general,Medico especialista,Enfermeria hospitalizacion'])->name('hospitalizacion.create');
+        Route::post('{id}', [HospitalizacionController::class, 'store'])
+            ->middleware(['role:Administrador,Medico general,Medico especialista,Enfermeria hospitalizacion'])->name('hospitalizacion.store');
+        Route::put('{id}', [HospitalizacionController::class, 'update'])
+            ->middleware(['role:Administrador,Medico general,Medico especialista,Enfermeria hospitalizacion'])->name('hospitalizacion.update');
+        Route::put('{id}', [HospitalizacionController::class, 'update'])
+            ->middleware(['role:Administrador,Medico general,Medico especialista,Enfermeria hospitalizacion'])->name('hospitalizacion.update');
+        Route::delete('{id}', [HospitalizacionController::class, 'destroy'])
+            ->middleware(['role:Administrador,Medico general,Medico especialista,Enfermeria hospitalizacion'])->name('hospitalizacion.eliminar');
+    });
+
+    Route::group(['prefix' => 'cart', 'middleware' => ['role:Administrador,Enfermeria hospitalizacion']], function () {
+    Route::post('add', [CartController::class, 'add'])->name('cart.add');
+    Route::post('add/enfermera', [CartController::class, 'addEnfermera'])->name('cart.addEnfermera');
+    Route::post('add/articuloHospi', [CartController::class, 'addArticuloHospi'])->name('cart.addArticuloHospi');
+    Route::post('add/articuloHonorario', [CartController::class, 'addArticuloHonorario'])->name('cart.addArticuloHonorario');    
+    Route::get('show/{hospitalizacion_id}', [CartController::class, 'show'])->name('cart.show');
+    Route::get('confirmar/{hospitalizacion_id}', [CartController::class, 'confirmar'])->name('cart.confirmar');
+    Route::put('{nombre} {hospitalizacion_id}', [CartController::class, 'update'])->name('cart.update');
+    });
+
+    Route::post('/guardar-hoja-hospital/{hospitalizacion_id}', [HojaHospitalController::class, 'store'])
+    ->middleware(['role:Administrador,Contador'])->name('guardarHojaHospital');
+
+
 });
 
 if (DB::table('users')->count() === 0) {
